debian download: use apt-get instead of aptitude
Whether aptitude is installed by default or not varies a lot by
distribution and distribution version and same with is it recommended or
discouraged.

apt-get is the only one which is surely in all Debian-based
distributions and versions by default. apt would work too, but it was
added in apt 1.0 and I don't think it was in oldoldstable which is still
supported.